Initialise the state vector from the seed rather than random memory
XorshiftRand::GenerateStates manipulates the state array, however the
calling code was not populating that array, only the array passed as
a prameter, so the state vector initialised with junk rather than
the seed.
This caused sector definitions to be unstable, and thus crashes when
calculating Faction distances when selecting a starting point.