Hacky global switch for Shader switching behaviour
If in GL2 mode...
Do Renderer::SetShaderSwitching(SWITCH_LAZY) to only switch shader
programs when a new program is required by a material, and never
switch to no shader program.
Do Renderer::SetSahderSwitching(SWITCH_SAFE) to revert to the the
current behaviour, which is to switch the shader program when
applying a material, and clear the program when unapplying a
material. Effectively every draw call that uses a material or
program will use and use that materials program.